[REFACTOR] divide code and modal refactor
