+++ React/.env
@@ -1,0 +1,5 @@
+REACT_APP_42_URL=https://api.intra.42.fr/oauth/authorize?client_id=u-s4t2ud-41b850ee7497100907e50e60530dea785e5dccfb80810e5c0852ff241c8c56d3&redirect_uri=http%3A%2F%2Flocalhost%3A3000%2Fauth%2Fcallback&response_type=code
+REACT_APP_HOST_IP=localhost
+REACT_APP_API_URL=http://${REACT_APP_HOST_IP}:4000
+REACT_APP_CLIENT_URL=http://${REACT_APP_HOST_IP}:3000
+REACT_APP_SOCKET_PORT=http://${REACT_APP_HOST_IP}:3001
