+++ React/src/App.tsx
@@ -7,8 +7,8 @@ import Verify from './Verify/Verify';
 import Main from './Main/Main';
 import Rank from './Rank/Rank';
 import Profile from './Profile/Profile';
-import OAuth from './Etc/OAuth';
-import NotFound from './Etc/NotFound';
+import OAuth from './Login/OAuth';
+import NotFound from './NotFound/NotFound';
 import Game from './Game/Game';
 import { CustomRoute } from './Util/CustomRoute';
 import './App.css';
