+++ React/src/Login/Login.tsx
@@ -1,21 +1,14 @@
-import { NavigateFunction, useNavigate } from 'react-router-dom';
-import CustomAxios from '../Util/CustomAxios';
+import { useNavigate } from 'react-router-dom';
+import loginClick from './loginClick';
 import './Login.css';
 
 function Login() {
-    const nav: NavigateFunction = useNavigate();
-    const loginClick = (): void => {
-        const token: string | null = localStorage.getItem('token');
-        if (token)
-            CustomAxios.get('/auth/jwt-verify').then(() => nav('/main'));
-        else
-            window.location.href = process.env.REACT_APP_42_URL || 'intra.42.fr';
-    }
+    const nav = useNavigate();
 
     return (
         <div id="login">
             <div id="title">Wally Pfister's<br/>PING 🏓 PONG</div>
-            <button id="login-button" onClick={loginClick}>Login</button>
+            <button id="login-button" onClick={()=>loginClick(nav)}>Login</button>
         </div>
     )
 }
