+++ React/src/Rank/Rank.tsx
@@ -1,27 +1,16 @@
-import RankUser from './RankUser';
-import './Rank.css'
 import { useQuery } from 'react-query';
-import CustomAxios from '../Util/CustomAxios';
-import NotFound from '../Etc/NotFound';
-import { StartGameModal, useStartGame } from '../Socket/StartGameModal';
-import { InviteFailModal, useInviteFail } from '../Socket/InviteFailedModal';
-import { InviteGameModal, useInviteGame } from '../Socket/InviteGameModal';
 import { useContext } from 'react';
 import { SocketContext } from '../Socket/SocketContext';
-
-const getRankData = async () => {
-    const res = await CustomAxios.get('/member/ranking');
-    return res.data;
-}
+import { SocketModalContainer } from '../Socket/SocketModal';
+import { UserProps } from './RankInterface';
+import NotFound from '../NotFound/NotFound';
+import RankUser from './RankUser';
+import getRankData from './getRankData';
+import './Rank.css'
 
 function Rank() {
     const socket = useContext(SocketContext);
-
     const { data, isLoading, isError } = useQuery('rank-data', getRankData, {retry: false, staleTime: 60 * 1000, refetchOnMount: 'always'});
-    const { showInvite, closeInvite, inviteData } = useInviteGame(socket);
-    const { showStart, closeStart, startData } = useStartGame(socket);
-    const { showInviteFail, closeInviteFail, inviteFailData } =
-      useInviteFail(socket);
 
     if (isLoading)
         return (<img src="../img/spinner.gif" alt="img"></img>);
@@ -30,6 +19,7 @@ function Rank() {
 
     return (
         <div id="rank">
+            <SocketModalContainer socket={socket}/>
             <div id="title">Ranking 👑</div>
             <div id="wrapper">
                 <div id="column">
@@ -41,33 +31,11 @@ function Rank() {
                 </div>
                 {
                     data.length === 0 ? <div id="no-user">No User</div> :
-                    data.map((user, idx) => (
+                    data.map((user: UserProps['user'], idx: number) => (
                         <RankUser user={user} idx={idx} key={idx}/>
                     ))
                 }
             </div>
-            {showInvite && (
-        <div className="invite-modal-overlay">
-          <InviteGameModal
-            onClose={closeInvite}
-            socket={socket}
-            inviteData={inviteData}
-          />
-        </div>
-      )}
-      {showStart && (
-        <div className="startgame-modal-overlay">
-          <StartGameModal onClose={closeStart} data={startData} />
-        </div>
-      )}
-      {showInviteFail && (
-        <div className="invite-fail-overlay">
-          <InviteFailModal
-            onClose={closeInviteFail}
-            inviteFailData={inviteFailData}
-          />
-        </div>
-      )}
         </div>
     )
 }
