+++ React/src/Rank/RankInterface.tsx
@@ -1,0 +1,10 @@
+export interface UserProps {
+    user: {
+        level: number,
+        name: string,
+        score: number,
+        win: number,
+        lose: number
+    },
+    idx: number
+}
