+++ React/src/Util/errorHandler.tsx
@@ -1,22 +1,17 @@
-import { AxiosError, InternalAxiosRequestConfig } from "axios";
+import { AxiosError } from "axios";
 import Swal from "sweetalert2";
 import CustomAxios from "./CustomAxios";
-
-interface ErrorResponse {
-    error?: string;
-    message?: string;
-    statusCode?: number;
-  }
+import { ErrorResponse } from "./UtilInterface";
 
 async function errorHandler(error: AxiosError) {
     if (error.response && error.response.status === 401) {
         const msg = (error.response.data as ErrorResponse).message;
         if (msg === "Invalid access token." || msg === "Unauthorized") {
             removeToken();
-            window.location.href = process.env.REACT_APP_CLIENT_URL || 'intra.42.fr';
+            window.location.href = process.env.REACT_APP_CLIENT_URL || 'https://intra.42.fr';
         } else if (msg === "Expired access token.") {
-            const originalRequest: InternalAxiosRequestConfig | undefined = error.config;
-            const refreshToken: string | null = localStorage.getItem('rtoken');
+            const originalRequest = error.config;
+            const refreshToken = localStorage.getItem('rtoken');
 
             if (originalRequest && refreshToken) {
                 await CustomAxios.get('/auth/jwt-refresh', {headers: {'Authorization': 'Bearer ' + refreshToken}}).then((res) => {
@@ -32,7 +27,7 @@ async function errorHandler(error: AxiosError) {
         removeToken();
         Swal.fire('관리자에게 문의해주세요.').then((res) => {
             if (res.isConfirmed) {
-                window.location.href = process.env.REACT_APP_CLIENT_URL || 'intra.42.fr';
+                window.location.href = process.env.REACT_APP_CLIENT_URL || 'https://intra.42.fr';
             }
         });
     }
