+++ React/src/Verify/Verify.tsx
@@ -1,33 +1,12 @@
-import LimitedAxios from '../Util/LimitedAxios';
-import Timer from '../Etc/Timer';
-import Swal from 'sweetalert2';
-import './Verify.css';
-import { NavigateFunction, useNavigate } from 'react-router-dom'
 import { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import Timer from '../Util/Timer';
+import verifyCode from './VerifyCode';
+import './Verify.css';
 
 function Verify() {
-    const nav: NavigateFunction = useNavigate();
     const [isRunning, setIsRunning] = useState(false);
-
-    const verifyCode = (event: React.MouseEvent<HTMLButtonElement>): void => {
-        event.preventDefault();
-
-        const target: HTMLButtonElement = event.target as HTMLButtonElement;
-        const input: HTMLInputElement = document.getElementById('verify-input') as HTMLInputElement;
-
-        target.setAttribute('disabled', '');
-        LimitedAxios.get('/auth/signin/tfa-verify', {params: {code: input.value}}).then((res) => {
-            const { accessToken, refreshToken } = res.data;
-            localStorage.setItem('token', accessToken);
-            localStorage.setItem('rtoken', refreshToken);
-            localStorage.removeItem('atoken');
-            nav('/main');
-        }).catch((err) => {
-            if (err.response.status === 409)
-                Swal.fire('Code is incorrect');
-            target.removeAttribute('disabled');
-        });
-    }
+    const nav = useNavigate();
     
     return (
         <div id="verify">
@@ -35,7 +14,7 @@ function Verify() {
             <Timer isRunning={isRunning} setIsRunning={setIsRunning}/>
             <div id="verify-form">
                 <input id="verify-input" placeholder="code"></input>
-                <button id="verify-button" onClick={verifyCode}>verify</button>
+                <button id="verify-button" onClick={(e)=>verifyCode(e, nav)}>verify</button>
             </div>
         </div>
     )
