+++ backend/.env
@@ -1,0 +1,47 @@
+# Environment variables declared in this file are automatically made available to Prisma.
+# See the documentation for more detail: https://pris.ly/d/prisma-schema#accessing-environment-variables-from-the-schema
+
+# Prisma supports the native connection string format for PostgreSQL, MySQL, SQLite, SQL Server, MongoDB and CockroachDB.
+# See the documentation for all the connection string options: https://pris.ly/d/connection-strings
+
+# local용
+# DATABASE_URL="postgresql://postgres:1234@localhost:5432/nestjs?schema=public"
+# docker용
+DATABASE_URL=postgresql://postgres:1234@postgres:5432/nestjs?schema=public
+
+# IP address
+HOST_IP_ADDRESS=localhost
+FRONT_PORT=http://${HOST_IP_ADDRESS}:3000
+SOCKET_PORT=http://${HOST_IP_ADDRESS}:3001
+
+# JWT
+JWT_ACCESS_SECRET=secretkeyforaccesstoken
+JWT_REFRESH_SECRET=secretkeyforrefreshtoken
+JWT_ACCESS_EXPIRE_TIME="1h" # One hour
+JWT_REFRESH_EXPIRE_TIME="1d" # One day
+
+# Limited JWT issued to 42 login users
+JWT_LIMITED_SECRET=secretkeyfortemptoken
+JWT_LIMITED_EXPIRE_TIME="10m" # Ten minutes
+
+# Temporary JWT issued to tfa verified users before signup
+JWT_SIGNUP_SECRET=secretkeyforsignuptoken
+JWT_SIGNUP_EXPIRE_TIME="10m" # Ten minutes
+
+# Session secret key
+SESSION_KEY=secretkeyforsession
+
+# 42 Oauth API
+FT_OAUTH_CLIENT_ID="u-s4t2ud-41b850ee7497100907e50e60530dea785e5dccfb80810e5c0852ff241c8c56d3"
+FT_OAUTH_SECRET="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"
+FT_OAUTH_CALLBACK=http://${HOST_IP_ADDRESS}:3000/auth/callback
+FT_OAUTH_URL=https://api.intra.42.fr/oauth/authorize
+FT_OAUTH_TOKEN_URL=https://api.intra.42.fr/oauth/token
+FT_API_ME=https://api.intra.42.fr/v2/me
+
+# Two-factor Authentication
+TFA_SEND_URL=http://${HOST_IP_ADDRESS}:3000/auth/tfa-send
+TFA_VERIFY_URL=http://${HOST_IP_ADDRESS}:3000/auth/tfa_verify
+TFA_EXPIRE_TIME="180" # Three minutes
+MAILER_NAME=tspong
+MAILER_PASS=blackhole519
