+++ backend/package.json
@@ -56,7 +56,7 @@
     "@nestjs/cli": "^9.0.0",
     "@nestjs/schematics": "^9.0.0",
     "@nestjs/testing": "^9.0.0",
-    "@prisma/client": "^4.14.0",
+    "@prisma/client": "^4.13.0",
     "@types/express": "^4.17.13",
     "@types/express-session": "^1.17.7",
     "@types/jest": "29.2.4",
@@ -76,7 +76,7 @@
     "nodemailer": "^6.9.1",
     "openssl": "^2.0.0",
     "prettier": "^2.3.2",
-    "prisma": "^4.14.0",
+    "prisma": "^4.13.0",
     "source-map-support": "^0.5.20",
     "supertest": "^6.1.3",
     "ts-jest": "29.0.3",
